feat: Rewrite media chat component
- Added ChatHeader, ChatInput, ChatMessageList, DragOverlay, FilePreview, ImageViewerDialog, and MediaChatComponent for enhanced chat functionality.
- Integrated file upload handling with drag-and-drop support.
- Implemented message rendering with support for text, images, videos, audio, and documents.
- Introduced scroll handling for loading older messages and maintaining scroll position.
- Updated type definitions for messages and message store.
